From: Raphael Kubo da Costa <kubo@profusion.mobi>
This signature change follows libcurl's behaviour more closely:
CURLOPT_POSTFIELDSIZE expects a long, and a value of -1 means that
content length calculation is forwarded to libcurl, which performs a
strlen() on CURLOPT_POSTFIELD.



SVN revision: 53845

+++ legacy/ecore/src/lib/ecore_con/Ecore_Con.h
@@ -508,7 +508,7 @@ EAPI Eina_Bool         ecore_con_url_httpauth_set(Ecore_Con_Url *url_con,
                                                   const char *password,
                                                   Eina_Bool safe);
 EAPI Eina_Bool         ecore_con_url_send(Ecore_Con_Url *url_con,
-                                          const void *data, size_t length,
+                                          const void *data, long length,
                                           const char *content_type);
 EAPI void              ecore_con_url_time(Ecore_Con_Url *url_con,
                                           Ecore_Con_Url_Time time_condition,
